sync

+++ include/utils/string.hpp
@@ -1,9 +1,8 @@
-#include <Types.hpp>
 extern "C" {
-void *memcpy(void *, const void *, size_t);
-void *memset(void *, int, size_t);
+void *memcpy(void *, const void *, unsigned long);
+void *memset(void *, int, unsigned long);
 int strcmp(const char *, const char *);
-int memcmp(const void *, const void *, size_t);
+int memcmp(const void *, const void *, unsigned long);
 unsigned strlen(const char *);
 long atol(const char *str);
 char *strchr(const char *str, int ch);
